requirements.txt
camunda-external-task-client-python3==3.2.0
certifi==2020.12.5
chardet==3.0.4
frozendict==1.2
idna==2.10
pika==1.2.0
pymongo==3.11.3
requests==2.23.0
urllib3==1.25.11
SQLAlchemy==1.3.23
psycopg2==2.8.6